requirements.txt
# Core Flask dependencies
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# Database
SQLAlchemy==2.0.38
alembic==1.14.1

# Authentication
google-auth==2.27.0
google-auth-oauthlib==1.2.0
oauthlib==2.1.0
requests-oauthlib==1.1.0

# AI/ML
anthropic==0.22.0
deepseek==1.0.0
pydantic==2.10.6
pydantic_core==2.27.2

# HTTP and networking
requests==2.32.3
urllib3==2.3.0
httpx==0.28.1
httpcore==1.0.7
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# Server
gunicorn==23.0.0

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
PyYAML==6.0.2

# Type checking and annotations
typing_extensions==4.12.2
annotated-types==0.7.0
